+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 
+ * Add `BuildExtension` helper for `maven-publish` and update zlib sample project
  * Add builds for Android to zlib sample project ([issue #5](https://github.com/bytedeco/gradle-javacpp/issues/5))
 
 ### September 9, 2020 version 1.5.4